refac (*) rename some features
