+++ processor/src/main/kotlin/kotli/template/multiplatform/compose/Rules.kt
@@ -39,7 +39,8 @@ object Rules {
     const val AppDIKt = "${CommonAppMainDir}/kotlin/kotli/app/di/DI.kt"
     const val AppScreenKt = "${CommonAppMainDir}/kotlin/kotli/app/AppScreen.kt"
     const val AppNavigationRouterKt = "${CommonAppMainDir}/kotlin/kotli/app/AppNavigationRouter.kt"
-    const val AppThemeDir = "${CommonAppMainDir}/kotlin/kotli/app/ui/theme"
+    const val AppUiDir = "${CommonAppMainDir}/kotlin/kotli/app/ui"
+    const val AppThemeDir = "${AppUiDir}/theme"
     const val AppThemeViewModelKt = "${AppThemeDir}/AppThemeViewModel.kt"
     const val AppThemeProviderKt = "${AppThemeDir}/AppThemeProvider.kt"
     const val AppThemeConfigDataKt = "${AppThemeDir}/AppThemeConfigData.kt"
@@ -47,7 +48,7 @@ object Rules {
     const val ProvidesNavigationBarStateKt = "${CommonAppMainDir}/kotlin/kotli/app/di/state/ProvidesNavigationBarState.kt"
     const val ProvidesNavigationStateKt = "${CommonAppMainDir}/kotlin/kotli/app/di/state/ProvidesNavigationState.kt"
     const val UserFlowDir = "${CommonAppMainDir}/kotlin/kotli/app/feature"
-    const val UserFlowDataLoaderDir = "${UserFlowDir}/loader/data"
+    const val UserFlowDataLoaderDir = "${AppUiDir}/loader"
     const val UserFlowThemeDir = "${UserFlowDir}/theme"
     const val UserFlowThemeChangeDir = "${UserFlowThemeDir}/change"
     const val UserFlowThemeToggleDir = "${UserFlowThemeDir}/toggle"
